feat: implement 'method implements interface' analyzer
The 'method implements interface' analyzer shows all interface methods that are implemented by the analyzed method.

The implementation is analogous to the 'interface method implemented by' analyzer (InterfaceMethodImplementedByNode.cs), that implements the reverse analysis direction of the analyzer implemented here in this commit.

In the current state, COM interfaces are not considered in this analyzer.